# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PM App API",
        "description": "Automated Project Management Tool API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/actions/": {
            "get": {
                "tags": ["actions"],
                "summary": "Read Actions",
                "description": "Get actions for a project.\n\nIf limit/offset are provided, returns paginated response with total count.\nOtherwise returns all actions (legacy behavior) with X-Total-Count header.\n\n- **limit**: Maximum number of results (1-100)\n- **offset**: Number of results to skip\n- **search**: Optional search term for title or Jira ID\n- **status**: Optional status filter (can be repeated)",
                "operationId": "read_actions_actions__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Project Id"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer",
                                    "maximum": 100,
                                    "minimum": 1
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Max items per page",
                            "title": "Limit"
                        },
                        "description": "Max items per page"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer",
                                    "minimum": 0
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Number of items to skip",
                            "title": "Offset"
                        },
                        "description": "Number of items to skip"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Search in title or Jira ID",
                            "title": "Search"
                        },
                        "description": "Search in title or Jira ID"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by status",
                            "title": "Status"
                        },
                        "description": "Filter by status"
                    },
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ActionItemRead"
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/PaginatedActionsResponse"
                                        }
                                    ],
                                    "title": "Response Read Actions Actions  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["actions"],
                "summary": "Create Action",
                "description": "Create a new action item.",
                "operationId": "create_action_actions__post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActionItemCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActionItemRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/actions/{action_id}": {
            "get": {
                "tags": ["actions"],
                "summary": "Get Action",
                "description": "Get a specific action item by ID.",
                "operationId": "get_action_actions__action_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "action_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Action Id"
                        }
                    },
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActionItemRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["actions"],
                "summary": "Delete Action",
                "description": "Delete an action item. Only Cogniters can delete actions.",
                "operationId": "delete_action_actions__action_id__delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "action_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Action Id"
                        }
                    },
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/actions/{action_id}/comments": {
            "get": {
                "tags": ["actions"],
                "summary": "Get Action Comments",
                "description": "Get all comments for an action item.\n\nBoth Cogniters and assigned Clients can view comments.",
                "operationId": "get_action_comments_actions__action_id__comments_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "action_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Action Id"
                        }
                    },
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CommentRead"
                                    },
                                    "title": "Response Get Action Comments Actions  Action Id  Comments Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["actions"],
                "summary": "Add Action Comment",
                "description": "Add a comment to an action item.\n\nBoth Cogniters and assigned Clients can comment on actions.",
                "operationId": "add_action_comment_actions__action_id__comments_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "action_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Action Id"
                        }
                    },
                    {
                        "name": "x-impersonate-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Impersonate-User-Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ActionItemCreate": {
                "properties": {
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ActionStatus",
                        "default": "To Do"
                    },
                    "assignee": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Assignee"
                    },
                    "due_date": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Due Date"
                    },
                    "priority": {
                        "$ref": "#/components/schemas/Priority",
                        "default": "Medium"
                    },
                    "jira_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Jira Id"
                    },
                    "jira_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Jira Key"
                    },
                    "issue_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Issue Type"
                    },
                    "project_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Project Id"
                    }
                },
                "type": "object",
                "required": ["title", "project_id"],
                "title": "ActionItemCreate",
                "description": "Schema for creating an action item."
            },
            "ActionItemRead": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "project_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Project Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ActionStatus"
                    },
                    "priority": {
                        "$ref": "#/components/schemas/Priority"
                    },
                    "comment_count": {
                        "type": "integer",
                        "title": "Comment Count"
                    },
                    "assignee": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Assignee"
                    },
                    "due_date": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Due Date"
                    },
                    "jira_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Jira Id"
                    },
                    "jira_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Jira Key"
                    },
                    "issue_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Issue Type"
                    },
                    "last_synced_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Synced At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "project_id",
                    "title",
                    "status",
                    "priority",
                    "comment_count",
                    "assignee",
                    "due_date",
                    "jira_id",
                    "jira_key",
                    "issue_type",
                    "last_synced_at"
                ],
                "title": "ActionItemRead",
                "description": "Schema for reading action item data.\n\nAll fields are explicitly defined to ensure OpenAPI required/nullable is accurate."
            },
            "ActionStatus": {
                "type": "string",
                "enum": ["To Do", "In Progress", "Complete", "No Status"],
                "title": "ActionStatus"
            },
...
            "CommentCreate": {
                "properties": {
                    "content": {
                        "type": "string",
                        "title": "Content"
                    }
                },
                "type": "object",
                "required": ["content"],
                "title": "CommentCreate",
                "description": "Schema for creating a comment."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PaginatedActionsResponse": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/ActionItemRead"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "limit": {
                        "type": "integer",
                        "title": "Limit"
                    },
                    "offset": {
                        "type": "integer",
                        "title": "Offset"
                    }
                },
                "type": "object",
                "required": ["items", "total", "limit", "offset"],
                "title": "PaginatedActionsResponse",
                "description": "Paginated response for action items."
            },
...
            "Priority": {
                "type": "string",
                "enum": ["High", "Medium", "Low"],
                "title": "Priority"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": ["loc", "msg", "type"],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "auth/login"
                    }
                }
            }
        }
    }
}